cleaning up and testing
